Add online canvas

+++ backend/src/routers/usersRouter.ts
@@ -1,6 +1,7 @@
 import {Elysia, t} from "elysia";
 import UserService from "../services/userService.ts";
 import {authMiddleware} from "../middlewares/authMiddleware.ts";
+import {ADMIN_USERNAME} from "../init.ts";
 
 export const usersRouter = new Elysia({
   prefix: '/api/users',
@@ -16,6 +17,7 @@ export const usersRouter = new Elysia({
     return {message: {users}}
   })
   .delete('/delete/:username', async ({user, status, params: {username}}) => {
+    if (username === ADMIN_USERNAME) return status('Forbidden', {error: 'Permission denied'});
     console.log(`Admin ${user.username} deleted ${username}.`)
     await UserService.removeUser(username)
     return status("OK")
@@ -25,6 +27,7 @@ export const usersRouter = new Elysia({
     })
   })
   .post('/update/', async ({status, body: {username, role}}) => {
+    if (username === ADMIN_USERNAME) return status('Forbidden', {error: 'Permission denied'});
     await UserService.updateUser(username, role)
     return status("OK")
   }, {
